pkg/idpool: split tests that timeout with race detector set
There is a test that times out in travis when running with the race
detector set, this is primarily caused by the high amount of go routines
started in parallel by such test. To avoid this issue we need to split
the test in 2 different files, one that has a lower maxID, and therefore
lower number of go routines started, for the builds with the race
detection, and another one with the same maxID for the builds without
the race detection.

Signed-off-by: André Martins <andre@cilium.io>