bpf: Hande icmpv6 in host firewall
In IPv6 mode, when the host firewall is enabled and rules are enforced,
we start dropping ICMPv6 packets that are required to route packets. In
particular, we can notice the following drops in cilium monitor:

  xx drop (Policy denied) flow 0x1fc0ef0 to endpoint 0, identity 1->0: fd00::c -> f00d::a0f:0:0:dfa1 DestinationUnreachable(NoRouteToDst)
  xx drop (Policy denied) flow 0x0 to endpoint 0, identity 0->0: fd01::c -> fd01::b NeighborAdvertisement

The nodes need to be able to exchange ICMPv6 NS and NA messages to
establish routes. We already handle the response to NS messages on
ingress, but when the egress policies are enforced, we start dropping
outgoing NS and NA messages.

This commit fixes that by allowing such messages on egress. No other
verifications are performed on the messages' correctness or their source
IP addresses. Such messages from the pods are already handled on their
egress, so we're not at risk of spoofing from pods here.

Fixes: 489dbef ("bpf: Enforce host policies for IPv6")
Signed-off-by: Paul Chaignon <paul@cilium.io>